Increase Timeout For AWS instances
The default waiting time is 10 minutes for waiting for compute instances to be launched. When launching large clusters this timeout will be hit just because of how long it takes for standup.  I am proposing increasing the default to 30 minutes to allow more time for the compute nodes to arrive.